pyproject.toml
[project]
name = "cs-streamlit-s"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.10",
    "httpx>=0.28.1",
    "langchain>=0.3.19",
    "langchain-cohere>=0.4.3",
    "langchain-community>=0.3.18",
    "langchain-core>=0.3.40",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.3.7",
    "langchain-qdrant>=0.2.0",
    "langchain-redis>=0.2.0",
    "langgraph>=0.3.2",
    "langsmith>=0.3.11",
    "openai>=1.65.2",
    "pydantic>=2.10.6",
    "pypdf>=5.3.1",
    "python-dotenv>=1.0.1",
    "qdrant-client<1.13.2",
    "redis>=5.2.1",
    "streamlit>=1.42.2",
    "tenacity>=9.0.0",
    "uvicorn>=0.34.0",
]
